# Added Performance class for collecting task timing info.