@frobnitzem 2021-03-10
There's also some minor issues I'll fix up:
  - the first 3 timers output in seconds, but should be like all the others (milliseconds)
  - the json formatting has some syntax errors