@psychocoderHPC 2021-03-11
@frobnitzem Thanks for your first contribution to PIConGPU. :heart: 

I will review the pull request.
I have a [prototype](https://github.com/psychocoderHPC/picongpu/tree/topic-flexP) with similar concepts as you implemented, the difference fo my prototype is that the profiling logic is in a plugin you can dynamically load by setting an environment variable.
I have not finished my prototype because we work on a more general solution: https://github.com/alpaka-group/bactria

I think your PR will definitively go into mainline and will hopefully also replace my hacky region timer from our Caar FOM branch.