@PrometheusPi 2021-03-11
@frobnitzem Great addition and thanks for your contribution!
Just a quick question: Is the performance output optional?