@psychocoderHPC 2021-03-30
@frobnitzem, @j-stephan, and me had a VC. We will integrate features of this PR in a [bacteria](https://github.com/alpaka-group/bactria) plugin. 
I try to integrate bacteria into PIConGPU in week 15.